chore: refactored App.tsx by splitting up Custom Sidebar, HeaderComponent and PageContent into separate component directories/files
