+++ src/components/custom-sidebar/children/nav-main.tsx
@@ -7,7 +7,7 @@ import {
 	SidebarMenu,
 	SidebarMenuButton,
 	SidebarMenuItem,
-} from '@/components/base-ui/sidebar.js';
+} from '@/components/base/sidebar.js';
 
 type AllowedPages =
 	| 'Hoare Logic Proof Validator'
