+++ src/components/custom-sidebar/children/settings.tsx
@@ -6,7 +6,7 @@ import {
 	SidebarMenu,
 	SidebarMenuButton,
 	SidebarMenuItem,
-} from '@/components/base-ui/sidebar.js';
+} from '@/components/base/sidebar.js';
 
 type AllowedPages =
 	| 'Hoare Logic Proof Validator'
